+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+* Hide new statistics screen in release variant. This was visible by fault and is not stable yet.
+
 ### Security
 
 ## [0.52.1] - 2026-08-30